Now gateway instance can be retrieved by packet-in's switch dpid or npt
