+++ src/main/java/net/floodlightcontroller/routing/IRoutingService.java
@@ -52,115 +52,6 @@ public interface IRoutingService extends IFloodlightService {
         }
     }
 
-
-    // L3 Routing Service below
-    /**
-     * Get all virtual gateway instances
-     * @return
-     */
-    Collection<VirtualGatewayInstance> getGatewayInstances();
-
-    /**
-     * Get a specific virtual gateway instance based on its name
-     * @param name
-     * @return
-     */
-    Optional<VirtualGatewayInstance> getGatewayInstance(String name);
-
-    /**
-     * Get all virtual interfaces associated with a given gateway
-     * @param gateway
-     * @return
-     */
-    Optional<Collection<VirtualGatewayInterface>> getGatewayInterfaces(VirtualGatewayInstance gateway);
-
-    /**
-     * Get a specific virtual interface with a given gateway
-     * @param name
-     * @param gateway
-     * @return
-     */
-    Optional<VirtualGatewayInterface> getGatewayInterface(String name, VirtualGatewayInstance gateway);
-
-    /**
-     * Add a virtual gateway instance
-     * @param gateway
-     */
-    void addGatewayInstance(VirtualGatewayInstance gateway);
-
-    /**
-     * Add a virtual interface onto gateway
-     * @param gateway
-     * @param intf
-     */
-    void addVirtualInterface(VirtualGatewayInstance gateway, VirtualGatewayInterface intf);
-
-    /**
-     * Update a specific virtual gateway instance
-     * @param name
-     * @param newMac
-     */
-    VirtualGatewayInstance updateVirtualGateway(String name, MacAddress newMac);
-
-    /**
-     * Update a specific virtual interface
-     * @param gateway
-     * @param intf
-     */
-    void updateVirtualInterface(VirtualGatewayInstance gateway, VirtualGatewayInterface intf);
-
-    /**
-     * Delete all existing virtual gateway instances
-     */
-    void deleteGatewayInstances();
-
-    /**
-     * Delete a specific virtual gateway instance
-     * @param name
-     * @return
-     */
-    boolean deleteGatewayInstance(String name);
-
-    /**
-     * Delete all virtual interfaces on gateway
-     * @param gateway
-     */
-    void removeAllVirtualInterfaces(VirtualGatewayInstance gateway);
-
-    /**
-     * Delete a specific virtual interface on gateway
-     * @param interfaceName
-     * @param gateway
-     * @return
-     */
-    boolean removeVirtualInterface(String interfaceName, VirtualGatewayInstance gateway);
-
-
-
-    Optional<Collection<VirtualSubnet>> getAllVirtualSubnets();
-    Optional<VirtualSubnet> getVirtualSubnet(String name);
-
-    SubnetMode getCurrentSubnetMode();
-
-    void createVirtualSubnet(String name, IPv4Address gatewayIP, DatapathId dpid);
-    void createVirtualSubnet(String name, IPv4Address gatewayIP, NodePortTuple npt);
-
-    boolean checkDPIDExist(DatapathId dpid);
-    boolean checkNPTExist(NodePortTuple nodePortTuple);
-
-    void updateVirtualSubnet(String name, IPv4Address gatewayIP, DatapathId dpid);
-    void updateVirtualSubnet(String name, IPv4Address gatewayIP, NodePortTuple npt);
-
-    void removeAllVirtualSubnets();
-    boolean removeVirtualSubnet(String name);
-
-//    boolean isSameSubnet(IPv4AddressWithMask ip1, IPv4AddressWithMask ip2);
-
-    boolean isSameSubnet(IOFSwitch sw1, IOFSwitch sw2);
-    boolean isSameSubnet(NodePortTuple npt1, NodePortTuple npt2);
-
-
-
     /**
      * Set the metric used when computing paths
      * across the topology.
